status resolver trabalhando apenas com minutos

+++ src/main/groovy/br/com/m2msolutions/copiloto/led/CopilotoLed.groovy
@@ -16,7 +16,7 @@ class CopilotoLed {
     private StatusResolver statusResolver = new StatusResolver()
     private DespachanteDeComando despachante
 
-    void enviarComando(Integer clienteId, Integer veiculoId, String modeloDoModulo, String moduloId,Double minutos,
+    void enviarComando(Integer clienteId, Integer veiculoId, String modeloDoModulo, String moduloId,Integer minutos,
                        Integer tempo = 1, Boolean openOrCloseTrip = false){
 
         def modulo = obterModulo modeloDoModulo
